pyproject.toml
[project]
name = "ws-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "KuznetsovAN",email = "kuznetsovan@uriit.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.45.0,<0.46.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "contextvars (>=2.4,<3.0)",
    "orjson (>=3.11.8,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2 (>=2.9.12,<3.0.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "websockets (>=16.0,<17.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "logger (>=1.4,<2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
